+++ visualize.py
@@ -261,8 +261,8 @@ def main(unused_argv):
                 predictions,
                 [0, 0, 0],
                 [1, original_image_shape[0], original_image_shape[1]])
-            resized_shape = tf.to_int32([tf.squeeze(samples[common.HEIGHT]),
-                                         tf.squeeze(samples[common.WIDTH])])
+            resized_shape = tf.cast([tf.squeeze(samples[common.HEIGHT]),
+                                         tf.squeeze(samples[common.WIDTH])], tf.int32)
             predictions = tf.squeeze(
                 tf.image.resize_images(tf.expand_dims(predictions, 3),
                                        resized_shape,
